+++ scripts/build-dist.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-set -ex
+set -Exu
+set -o pipefail
 # Builds the release tarball for Yarn.
 
 umask 0022 # Ensure permissions are correct (0755 for dirs, 0644 for files)
@@ -16,8 +17,9 @@ case "$(uname -s)" in
     ;;
 esac
 
+version=`node -e "console.log(JSON.parse(fs.readFileSync('package.json')).version)"`
+
 rm -rf artifacts dist
-rm -rf dist
 mkdir artifacts
 mkdir dist{,/bin,/lib}
 
@@ -25,16 +27,23 @@ mkdir dist{,/bin,/lib}
 eval $system_yarn run build
 eval $system_yarn run build-bundle
 chmod +x artifacts/*.js
+# Verify that it works as expected
+[[ "$version" == $(node artifacts/yarn-legacy-$version.js --version) ]] || exit 1
+[[ "$version" == "$(node artifacts/yarn-$version.js --version)" ]] || exit 1
 
 cp package.json dist/
 cp README.md dist/
 cp LICENSE dist/
-cp artifacts/yarn-legacy-*.js dist/lib/yarn-cli.js
-cp bin/yarn-bundle-entry.js dist/bin/yarn.js
-cp bin/{yarn,yarnpkg,*.cmd} dist/bin/
+cp artifacts/yarn-legacy-$version.js dist/lib/yarn-legacy.js
+cp artifacts/yarn-$version.js dist/lib/yarn.js
+cp bin/{yarn.js,yarn,yarnpkg,*.cmd} dist/bin/
+chmod +x dist/bin/*
+
 # We cannot bundle v8-compile-cache as it must be loaded separately to be effective.
 cp node_modules/v8-compile-cache/v8-compile-cache.js dist/lib/v8-compile-cache.js
 
-version=`exec $dist_yarn --version`
+# Verify that it works as expected
+[[ "$version" == "$(./dist/bin/yarn --version)" ]] || exit 1;
+
 ./scripts/update-dist-manifest.js $(readlink -f dist/package.json) tar
 tar -cvzf artifacts/yarn-v$version.tar.gz dist/*
